Move App subcomponents to helpers
